+++ executor_nocoverage.go
@@ -1,0 +1,3 @@
+package dogma
+
+func (EventObserverOption) isExecuteCommandOption() {}