+++ CHANGELOG.md
@@ -12,12 +12,12 @@ The format is based on [Keep a Changelog], and this project adheres to
 [bc]: https://github.com/dogmatiq/.github/blob/main/VERSIONING.md#changelogs
 [engine bc]: https://github.com/dogmatiq/.github/blob/main/VERSIONING.md#changelogs
 
-## [Unreleased]
+## [0.19.1] - 2025-12-09
 
 ### Added
 
-- Added `WithEventObserver()` to support observing events recorded by a specific
-  invocation of `ExecuteCommand()`.
+- **[ENGINE BC]** Added `WithEventObserver()` to support observing events
+  recorded by a specific invocation of `ExecuteCommand()`.
 
 ## [0.19.0] - 2025-12-08
 
@@ -526,6 +526,7 @@ No engines except [testkit] are able to provide a meaningful implementation of
 [0.18.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.18.0
 [0.18.1]: https://github.com/dogmatiq/dogma/releases/tag/v0.18.1
 [0.19.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.19.0
+[0.19.1]: https://github.com/dogmatiq/dogma/releases/tag/v0.19.1
 
 <!-- adr references -->